flattened abstraction
